+++ sources/Messages/PlayerInfo.cs
@@ -6,8 +6,8 @@
     {
         public DateTime Time { get; set; }
         public int Id { get; set; }
-        public int HitsPercents { get; set; }
-        public int BattleAvgXp { get; set; }
+        public double HitsPercents { get; set; }
+        public double BattleAvgXp { get; set; }
         public int Draws { get; set; }
         public int Wins { get; set; }
         public int Losses { get; set; }
